Agora a medida de conversao da unidade de medida aceita double

+++ src/br/ages/crud/command/AddUnidadeMedidaCommand.java
@@ -34,7 +34,7 @@ public class AddUnidadeMedidaCommand implements Command {
 		String descricaoOrigem = request.getParameter("descricaoOrigem");
 		String descricaoConversao = request.getParameter("descricaoConversao");
 		String sigla = request.getParameter("sigla");
-		int medidaConversao = Integer.valueOf(request.getParameter("medidaConversao"));
+		double medidaConversao = Double.valueOf(request.getParameter("medidaConversao"));
 		
 		try {
 			UnidadeMedida unidadeMedida = new UnidadeMedida();
